Address the PR review of the Boost.Asio block sink
- Move the duplicated `AD_CONTRACT_CHECK(chunkIndex < numChunks_)` from
  `asyncPush` and `asyncFinishChunk` into their shared `sendToChunk`.
- Stop the merge only on the *first* pushed exception. A later one has nothing
  left to wake, because `requestStop` has already woken every suspended
  operation and no channel is ever created after the stop.
- Explain at `asyncGetNextBlock` why the single-consumer requirement is not
  subsumed by the strand: the strand serializes the individual steps of an
  operation and not two whole operations, so two concurrent calls would lose
  the global order and one of them would erase the channel that the other is
  still waiting on.
- Make `completeOn` use `net::post` instead of `net::dispatch`, so that a
  completion handler of the sink never runs while the strand is held.
  `net::dispatch` invokes the handler inline whenever the calling thread
  already belongs to the target executor, which is exactly the case for a
  producer whose handler is bound to the thread pool that also runs the strand;
  such a producer then merged a whole output block with the strand held and
  thereby serialized the entire merge. With that guarantee in the sink, a
  `ChunkTask` simply binds its handlers to `executor_` and continues merging
  directly in the handler, so `ChunkTask::postStep` is no longer needed.
- Reuse `ad_utility::runFunctionOnExecutor` instead of the hand-written
  `postToStrand`. To keep the sink free of coroutines, that helper moves from
  `util/AsioHelpers.h` (which pulls in coroutines and `beast`) into the new
  `util/RunFunctionOnExecutor.h`; it also now takes its completion token by
  forwarding reference, such that rvalue tokens bind, and no longer expands a
  parameter pack in a lambda init-capture, such that it works in C++17.
